tc_networkthroughput2: Use productType of correct host in heterogeneous tests
We need to look at the productType of the host to determine how to find out various things. But sometimes the hosts involved in the test may be of different types. So we cannot rely on self.host being appropriate for all hosts in the test. Instead, we must get a handle on the appropriate Host object first.