Allow PRODUCT_TYPE to override default assumption that host is XenServer
This allows us to not require a hypervisor-specific sequence file explicitly stating productType="esx" (or similar) -- we can just set the PRODUCT_TYPE variable.

(This commit is a corrected version of c/s 426ae293cbf, which was reverted. The difference is that we are no longer universally setting work["productVersion"] to the value of the PRODUCT_VERSION variable in handleHostNode, because for XenServer we seem to rely on this being None. If it's not None, the first thing exec/xenrt/lib/xenserver/host.py's createHost() function tries to do is call productInputdirForVersion(productVersion), which fails for strings like "Sarasota". So now we only set work["productVersion"] to the value of PRODUCT_VERSION for esx hosts, where this seems to be useful.)