Use SQL parameters to avoid escaping issues
There are a number of cases where we are not escaping strings, and thus causing DB insert failures. Rather than attempting to use our own possibly broken function, use the support built in to pgdb.

This also avoids some SQL injection issues, though the threat model for XenRT has not to this point had that as an issue anyway...