RHEL7: flush all firewall rules to avoid blocking synexec, iperf, etc.
Previously, I had tried to manipulate the RHEL7 firewall rules to allow traffic
from specific services. But in that approach it is too cumbersome to add new
exceptions. Instead, let's get out the bigger hammer and flush all the rules.

Ideally, this kind of thing would be done in the %post section of the
installation script, but I can't work out how to flush the rules permanently,
hence it's done on each boot via /etc/rc.local.