Revert "Use correct generic Debian template in Tampa, Clearwater, Sarasota"
This reverts commit 0e24c526949379c5ed6531b3a0b45c88808e4119.

It looks like this patch tries to add Squeeze to the list of templates used for
"TEMPLATE_NAME_DEBIAN".  I believe "TEMPLATE_NAME_DEBIAN" is only used when
distro=="debian".  I think the only time distro=="debian" is in
createGenericLinuxGuest, for which we carry a patch that makes it use
generic_distro rather than "debian".  So this patch is believed redundant.