CP-10044: fix bootstorm issue with esx on libvirt NFSStorage SR.create
libvirt does not accept the nosubdir parameter in the NFSStorage SR.create method
that is called by sequence.py, raising the following during a bootstorm using an NFS SR:

create() got an unexpected keyword argument 'nosubdir'Traceback (most recent call last):
  File "/usr/share/xenrt/venvs/878365/exec/xenrt/sequence.py", line 1716, in runThis
    sr.create(server, path, nosubdir=nosubdir)
TypeError: create() got an unexpected keyword argument 'nosubdir'
[VERBOSE] [2014-10-27 11:40:14 UTC] tryupto: exception create() got an unexpected keyword argument 'nosubdir'

nosubdir=True seems only useful in the SMAPI in XenServer.
nosubdir=False is the default in sequence.py

This change only accepts the new parameter in the method, and does nothing with it,
defaulting to the behaviour when nosubdir=False.