ESXi: implement cloneVDI using linked thin-clones
An implementation of cloning using copy-on-write thin-clones by exploiting
ESXi's snapshot functionality.

Previously, cloneVDI fell back to a full-fat copy. Now, the behaviour of
cloneVDI is in line with the usually expected "fast-clone" behaviour, as
implemented in the XenServer library.

At present, the implementation assumes that the VM being cloned only has one
disk.