Add a function 'getAssumedId' to Host classes
This function returns the assumedid of the network matching a given "friendly
network name". It assumes there is only one NIC having that name, much like
getNetworkUUID does in xenserver.Host.

Previously, this functionality was implemented in tc_netthru2 and switch-cased
on the type of host. It is much more natural to have this as a method in the
Host classes.