Check the size as reported by HEAD of files in the shared cache
We have had several situations where users have changed files 'underneath' XenRT, which has caused issues when XenRT finds the old version in its cache.
This commit modifies the FileManager so it attempts an HTTP HEAD request and compares the content-length reported with the size of file it has.

If there is no content-length or it's less than 10M we don't validate on the basis it's probably a script or similar serving the file.

Another option in future might be to store and then check the last modified dates, but this is probably sufficient for the obvious cases.