Default job email to the user's email address
If no email parameter is specified this will now default to the user's e-mail address. To avoid this (i.e. to specify no email) a null parameter or empty string can be provided.
Also modify the CLI to send email using the top level field rather than as an entry in the params dictionary