Check if custom sequences / testcase files actually exist
At the moment if you specify a custom sequence which doesn't exist, the job gets submitted then a curl error occurs when the upload is attempted.
This patch adds checking at the point of parsing the arguments to verify the custom sequence / testcase files exist, and raises an exception at that point (i.e. before the job has been submitted) if they aren't