Fix an ACL double counting bug
This commit fixes an issue wherby when a user asks to borrow a machine they're already running a job on (either directly or as part of the job), they would only be allowed to if they would be allowed an additional machine under the ACL.
To resolve this the ACL checker needs to know the names of the machines the user is requesting, so it can ignore any they already have, as such all calls and existing unit tests have been updated to do this. In the scheduler this is slightly cheating as we just pass the first number machines from the list through, however the only case we need to care about this is if the user already has the machines borrowed, and in that case code inspection shows the correct behaviour will occur.
Also add new unit tests for the double counting bug