lib/netscaler: support tagged VLAN on SR-IOV VF
This is very fragile, but I can't find another way of making this work.

You would think that this would work:
 * add vlan 3601
 * bind vlan 3601 -ifnum 10/1 -tagged

But this results in the error "ERROR: Maximum number of tagged VLANs bound to
the interface exceeded or the binding of this VLAN is not allowed on the
interface."

This leads us to try the following:
 * set interface 10/1 -tagall ON
 * add vlan 3601
 * bind vlan 3601 -ifnum 10/1

This appears to work, but when you save the config and reboot, "show vlan" shows
that it hasn't persisted the configuration.

Instead, we need to do a magic trick:
 * set interface 10/1 -tagall OFF
 * add vlan 3601
 * bind vlan 3601 -ifnum 10/1
 * set interface 10/1 -tagall ON

... and don't reboot!

Hence anything relying on tagged VLANs must not cause a reboot of the VPX after
the network configuration.