tc_diskconc2: don't assume guest disk is at /dev/xvd[a-z]
For example, on ESXi, the guest disk is at /dev/sd[a-z]. So let's use the same
trick we used for latency, i.e. /dev/*d[a-z].

Note that we need to use the sub-shell to get the name of this device because
the shell wouldn't extrapolate the * due to the leading '--filename='.