Sync RPMs etc using sudo
Some Linux RPMs come from ISOs, at which point the permissions on the files are often such that they prevent the owner from writing. This means in the situation where a partial sync has occurred, rsync can't "finish the job" as it may not have write permission to the subdirectories.

By using sudo we can workaround this limitation