CA-188697 Fix checkGuestMemory and move guest memory logic to OS classes
This change creates a new property 'visibleMemory' on the OS interface, which returns the amount of memory visible to the guest in MB.
The existing method is now just a wrapper for the new method.
Also remove all concept of unit (and just define it to be in MB) from memory methods.