Only schedule a machines specified job if all the machines are in the same site
At the moment you can submit a job specifying machines across sites - while they are all allocated to the job, those machines not in the same site as the first machine will not be useable by XenRT, thus leading to confusing results.

This change therefore modifies the behaviour so we reject any specified machines in a different site to the first one.